update to updateBase.select
updated select function to check for the existence of the passed attributeKey before retrieving value, otherwise uses the key itself as the selector.  This allows you to still call this.select('someSelector'), but also allows for a call of  this.select(this.attr.someSelector) which I find somewhat more intuitive.